+++ pages/pagevescpackage.cpp
@@ -57,6 +57,23 @@ PageVescPackage::PageVescPackage(QWidget *parent) :
     on_outputRefreshButton_clicked();
 
     reloadArchive();
+
+    mDescriptionUpdated = true;
+    connect(ui->descriptionEdit, &QMarkdownTextEdit::textChanged, [this]() {
+        mDescriptionUpdated = true;
+    });
+
+    mPreviewTimer = new QTimer(this);
+    mPreviewTimer->start(500);
+    connect(mPreviewTimer, &QTimer::timeout, [this]() {
+        if (mDescriptionUpdated) {
+            mDescriptionUpdated = false;
+            auto posOld = ui->descriptionBrowser->verticalScrollBar()->value();
+            ui->descriptionBrowser->setHtml(
+                        Utility::md2html(ui->descriptionEdit->document()->toPlainText()));
+            ui->descriptionBrowser->verticalScrollBar()->setValue(posOld);
+        }
+    });
 }
 
 PageVescPackage::~PageVescPackage()
@@ -127,8 +144,27 @@ void PageVescPackage::on_chooseOutputButton_clicked()
             filename += ".vescpkg";
         }
 
-        ui->outputEdit->setText(filename);
-        on_outputRefreshButton_clicked();
+        if (QFile::exists(filename)) {
+            if (ui->descriptionEdit->toPlainText().size() > 10) {
+                QMessageBox::StandardButton reply =
+                        QMessageBox::warning(this,
+                                             tr("Replace Content"),
+                                             tr("Opening an existing package will replace the content "
+                                                "in the editor. Do you want to continue?"),
+                                             QMessageBox::Ok | QMessageBox::Cancel);
+
+                if (reply == QMessageBox::Ok) {
+                    ui->outputEdit->setText(filename);
+                    on_outputRefreshButton_clicked();
+                }
+            } else {
+                ui->outputEdit->setText(filename);
+                on_outputRefreshButton_clicked();
+            }
+        } else {
+            ui->outputEdit->setText(filename);
+            on_saveButton_clicked();
+        }
     }
 }
 
@@ -182,14 +218,12 @@ void PageVescPackage::on_saveButton_clicked()
     }
 
     pkg.name = ui->nameEdit->text();
-    pkg.description = ui->descriptionEdit->document()->toHtml();
+    pkg.description = ui->descriptionEdit->document()->toPlainText();
 
     file.write(mLoader.packVescPackage(pkg));
     file.close();
 
-    mVesc->emitMessageDialog(tr("Save Package"),
-                             tr("Package Saved"),
-                             true, false);
+    mVesc->emitStatusMessage(tr("Package Saved"), true);
 }
 
 void PageVescPackage::on_loadRefreshButton_clicked()
@@ -200,7 +234,13 @@ void PageVescPackage::on_loadRefreshButton_clicked()
     }
 
     auto pkg = mLoader.unpackVescPackage(f.readAll());
-    ui->loadBrowser->document()->setHtml(pkg.description);
+
+    QString line1 = QTextStream(&pkg.description).readLine();
+    if (line1.contains("<!DOCTYPE HTML PUBLIC", Qt::CaseInsensitive)) {
+        ui->loadBrowser->document()->setHtml(pkg.description);
+    } else {
+        ui->loadBrowser->document()->setHtml(Utility::md2html(pkg.description));
+    }
 }
 
 void PageVescPackage::on_writeButton_clicked()
@@ -229,7 +269,18 @@ void PageVescPackage::on_outputRefreshButton_clicked()
     }
 
     auto pkg = mLoader.unpackVescPackage(f.readAll());
-    ui->descriptionEdit->document()->setHtml(pkg.description);
+
+    QString line1 = QTextStream(&pkg.description).readLine();
+    if (line1.contains("<!DOCTYPE HTML PUBLIC", Qt::CaseInsensitive)) {
+        ui->descriptionEdit->document()->setHtml(pkg.description);
+        QString md = ui->descriptionEdit->document()->toMarkdown();
+        ui->descriptionEdit->document()->setPlainText(md);
+    } else {
+        ui->descriptionEdit->document()->setPlainText(pkg.description);
+    }
+
+    mDescriptionUpdated = true;
+
     ui->nameEdit->setText(pkg.name);
 }
 
@@ -319,7 +370,7 @@ void PageVescPackage::reloadArchive()
     ui->applicationList->clear();
     ui->libraryList->clear();
 
-    for (auto p: pList) {
+    foreach (auto p, pList) {
         auto pVal = p.value<VescPackage>();
         QListWidgetItem *item = new QListWidgetItem;
         item->setText(pVal.name);
@@ -336,7 +387,7 @@ void PageVescPackage::reloadArchive()
 void PageVescPackage::packageSelected(VescPackage pkg)
 {
     mCurrentPkg = pkg;
-    ui->storeBrowser->document()->setHtml(pkg.description);
+    ui->storeBrowser->document()->setHtml(Utility::md2html(pkg.description));
     ui->installButton->setEnabled(!pkg.isLibrary);
     if (ui->installButton->isEnabled()) {
         ui->installButton->setToolTip("");